// PROJECT/src/model/order.mode.ts
export interface Order{
    Company_Name: string;
    Represented_Name: string;
    Email: string;
    Contact_No: string;
    Street_Address: string;
    City: string;
    Postal_ZipCode: string;
    State: string;
    Country: string;
    Commodity_Product_Oil: string;
    Quantity: string;
    MT_Barrel_Gallon: string;
    Term_Delivery: string;
   }